Simple Wordle Complete
Created a simple version of wordle

User has three chances to guess 4 letter word. Each guess has an editable text box that the user can enter their guess into. Text box only takes alphabet input, has a maximum possible letter count of 4, and the user is unable to submit a guess if it is less than 4 letters

After each guess, a new textbox appears for the user to input their guess while the previous one is greyed out and disabled, only displaying how close to the previous the player was

After 3 guesses, the submit button goes away and is replaced with the reset button, which returns the game to its initial state with a new word to guess

The word to be guessed is displayed after 3 wrong guesses or whenever the player gets the correct word, whichever comes first